requirements.txt
amqp==5.3.1
asgiref==3.11.0
asn1crypto==1.5.1
billiard==4.2.4
celery==5.3.6
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.8
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==25.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.95.0
idna==3.11
joblib==1.5.3
kombu==5.6.1
nltk==3.8.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
regex==2025.11.3
reportlab==4.0.7
requests==2.32.5
scramp==1.4.6
six==1.17.0
sorl-thumbnail==12.11.0
sqlparse==0.5.5
stripe==14.1.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
uvicorn
gunicorn
dj-database-url
whitenoise
